Split up the request modal into modules for each dialog variant
